产品管理多租户改造,激活设备多租户改造

+++ controllers/tp_product.go
@@ -22,111 +22,131 @@ type TpProductController struct {
 }
 
 // 列表
-func (TpProductController *TpProductController) List() {
-	PaginationValidate := valid.TpProductPaginationValidate{}
-	err := json.Unmarshal(TpProductController.Ctx.Input.RequestBody, &PaginationValidate)
+func (c *TpProductController) List() {
+	reqData := valid.TpProductPaginationValidate{}
+	err := json.Unmarshal(c.Ctx.Input.RequestBody, &reqData)
 	if err != nil {
 		fmt.Println("参数解析失败", err.Error())
 	}
 	v := validation.Validation{}
-	status, _ := v.Valid(PaginationValidate)
+	status, _ := v.Valid(reqData)
 	if !status {
 		for _, err := range v.Errors {
 			// 获取字段别称
-			alias := gvalid.GetAlias(PaginationValidate, err.Field)
+			alias := gvalid.GetAlias(reqData, err.Field)
 			message := strings.Replace(err.Message, err.Field, alias, 1)
-			response.SuccessWithMessage(1000, message, (*context2.Context)(TpProductController.Ctx))
+			response.SuccessWithMessage(1000, message, (*context2.Context)(c.Ctx))
 			break
 		}
 		return
 	}
+	//获取租户id
+	tenantId, ok := c.Ctx.Input.GetData("tenant_id").(string)
+	if !ok {
+		response.SuccessWithMessage(400, "代码逻辑错误", (*context2.Context)(c.Ctx))
+		return
+	}
 	var TpProductService services.TpProductService
-	isSuccess, d, t := TpProductService.GetTpProductList(PaginationValidate)
+	isSuccess, d, t := TpProductService.GetTpProductList(reqData, tenantId)
 
 	if !isSuccess {
-		response.SuccessWithMessage(1000, "查询失败", (*context2.Context)(TpProductController.Ctx))
+		response.SuccessWithMessage(1000, "查询失败", (*context2.Context)(c.Ctx))
 		return
 	}
 	dd := valid.RspTpProductPaginationValidate{
-		CurrentPage: PaginationValidate.CurrentPage,
+		CurrentPage: reqData.CurrentPage,
 		Data:        d,
 		Total:       t,
-		PerPage:     PaginationValidate.PerPage,
+		PerPage:     reqData.PerPage,
 	}
-	response.SuccessWithDetailed(200, "success", dd, map[string]string{}, (*context2.Context)(TpProductController.Ctx))
+	response.SuccessWithDetailed(200, "success", dd, map[string]string{}, (*context2.Context)(c.Ctx))
 }
 
 // 编辑
-func (TpProductController *TpProductController) Edit() {
-	TpProductValidate := valid.TpProductValidate{}
-	err := json.Unmarshal(TpProductController.Ctx.Input.RequestBody, &TpProductValidate)
+func (c *TpProductController) Edit() {
+	reqData := valid.TpProductValidate{}
+	err := json.Unmarshal(c.Ctx.Input.RequestBody, &reqData)
 	if err != nil {
 		fmt.Println("参数解析失败", err.Error())
 	}
 	v := validation.Validation{}
-	status, _ := v.Valid(TpProductValidate)
+	status, _ := v.Valid(reqData)
 	if !status {
 		for _, err := range v.Errors {
 			// 获取字段别称
-			alias := gvalid.GetAlias(TpProductValidate, err.Field)
+			alias := gvalid.GetAlias(reqData, err.Field)
 			message := strings.Replace(err.Message, err.Field, alias, 1)
-			response.SuccessWithMessage(1000, message, (*context2.Context)(TpProductController.Ctx))
+			response.SuccessWithMessage(1000, message, (*context2.Context)(c.Ctx))
 			break
 		}
 		return
 	}
-	if TpProductValidate.Id == "" {
-		response.SuccessWithMessage(1000, "id不能为空", (*context2.Context)(TpProductController.Ctx))
+	if reqData.Id == "" {
+		response.SuccessWithMessage(1000, "id不能为空", (*context2.Context)(c.Ctx))
+		return
+	}
+	//获取租户id
+	tenantId, ok := c.Ctx.Input.GetData("tenant_id").(string)
+	if !ok {
+		response.SuccessWithMessage(400, "代码逻辑错误", (*context2.Context)(c.Ctx))
+		return
 	}
 	var TpProductService services.TpProductService
-	isSucess := TpProductService.EditTpProduct(TpProductValidate)
+	isSucess := TpProductService.EditTpProduct(reqData, tenantId)
 	if isSucess {
-		d := TpProductService.GetTpProductDetail(TpProductValidate.Id)
-		response.SuccessWithDetailed(200, "success", d, map[string]string{}, (*context2.Context)(TpProductController.Ctx))
+		d := TpProductService.GetTpProductDetail(reqData.Id)
+		response.SuccessWithDetailed(200, "success", d, map[string]string{}, (*context2.Context)(c.Ctx))
 	} else {
-		response.SuccessWithMessage(400, "编辑失败", (*context2.Context)(TpProductController.Ctx))
+		response.SuccessWithMessage(400, "编辑失败", (*context2.Context)(c.Ctx))
 	}
 }
 
 // 新增
-func (TpProductController *TpProductController) Add() {
-	AddTpProductValidate := valid.AddTpProductValidate{}
-	err := json.Unmarshal(TpProductController.Ctx.Input.RequestBody, &AddTpProductValidate)
+func (c *TpProductController) Add() {
+	reqData := valid.AddTpProductValidate{}
+	err := json.Unmarshal(c.Ctx.Input.RequestBody, &reqData)
 	if err != nil {
 		fmt.Println("参数解析失败", err.Error())
 	}
 	v := validation.Validation{}
-	status, _ := v.Valid(AddTpProductValidate)
+	status, _ := v.Valid(reqData)
 	if !status {
 		for _, err := range v.Errors {
 			// 获取字段别称
-			alias := gvalid.GetAlias(AddTpProductValidate, err.Field)
+			alias := gvalid.GetAlias(v, err.Field)
 			message := strings.Replace(err.Message, err.Field, alias, 1)
-			response.SuccessWithMessage(1000, message, (*context2.Context)(TpProductController.Ctx))
+			response.SuccessWithMessage(1000, message, (*context2.Context)(c.Ctx))
 			break
 		}
 		return
 	}
+	//获取租户id
+	tenantId, ok := c.Ctx.Input.GetData("tenant_id").(string)
+	if !ok {
+		response.SuccessWithMessage(400, "代码逻辑错误", (*context2.Context)(c.Ctx))
+		return
+	}
 	var TpProductService services.TpProductService
-	if AddTpProductValidate.Plugin == "" {
-		AddTpProductValidate.Plugin = "{}"
+	if reqData.Plugin == "" {
+		reqData.Plugin = "{}"
 	}
 	id := uuid.GetUuid()
 	TpProduct := models.TpProduct{
 		Id:            id,
-		ProtocolType:  AddTpProductValidate.ProtocolType,
-		AuthType:      AddTpProductValidate.AuthType,
-		Describe:      AddTpProductValidate.Describe,
+		ProtocolType:  reqData.ProtocolType,
+		AuthType:      reqData.AuthType,
+		Describe:      reqData.Describe,
 		CreatedTime:   time.Now().Unix(),
-		Name:          AddTpProductValidate.Name,
-		Plugin:        AddTpProductValidate.Plugin,
-		Remark:        AddTpProductValidate.Remark,
-		SerialNumber:  AddTpProductValidate.SerialNumber,
-		DeviceModelId: AddTpProductValidate.DeviceModelId,
+		Name:          reqData.Name,
+		Plugin:        reqData.Plugin,
+		Remark:        reqData.Remark,
+		SerialNumber:  reqData.SerialNumber,
+		DeviceModelId: reqData.DeviceModelId,
+		TenantId:      tenantId,
 	}
 	rsp_err, d := TpProductService.AddTpProduct(TpProduct)
 	if rsp_err == nil {
-		response.SuccessWithDetailed(200, "success", d, map[string]string{}, (*context2.Context)(TpProductController.Ctx))
+		response.SuccessWithDetailed(200, "success", d, map[string]string{}, (*context2.Context)(c.Ctx))
 	} else {
 		var err string
 		isTrue := strings.Contains(rsp_err.Error(), "23505")
@@ -135,39 +155,47 @@ func (TpProductController *TpProductController) Add() {
 		} else {
 			err = rsp_err.Error()
 		}
-		response.SuccessWithMessage(400, err, (*context2.Context)(TpProductController.Ctx))
+		response.SuccessWithMessage(400, err, (*context2.Context)(c.Ctx))
 	}
 }
 
 // 删除
-func (TpProductController *TpProductController) Delete() {
-	TpProductIdValidate := valid.TpProductIdValidate{}
-	err := json.Unmarshal(TpProductController.Ctx.Input.RequestBody, &TpProductIdValidate)
+func (c *TpProductController) Delete() {
+	reqData := valid.TpProductIdValidate{}
+	err := json.Unmarshal(c.Ctx.Input.RequestBody, &reqData)
 	if err != nil {
 		fmt.Println("参数解析失败", err.Error())
 	}
 	v := validation.Validation{}
-	status, _ := v.Valid(TpProductIdValidate)
+	status, _ := v.Valid(reqData)
 	if !status {
 		for _, err := range v.Errors {
 			// 获取字段别称
-			alias := gvalid.GetAlias(TpProductIdValidate, err.Field)
+			alias := gvalid.GetAlias(reqData, err.Field)
 			message := strings.Replace(err.Message, err.Field, alias, 1)
-			response.SuccessWithMessage(1000, message, (*context2.Context)(TpProductController.Ctx))
+			response.SuccessWithMessage(1000, message, (*context2.Context)(c.Ctx))
 			break
 		}
 		return
 	}
-	if TpProductIdValidate.Id == "" {
-		response.SuccessWithMessage(1000, "id不能为空", (*context2.Context)(TpProductController.Ctx))
+	if reqData.Id == "" {
+		response.SuccessWithMessage(1000, "id不能为空", (*context2.Context)(c.Ctx))
+		return
+	}
+	//获取租户id
+	tenantId, ok := c.Ctx.Input.GetData("tenant_id").(string)
+	if !ok {
+		response.SuccessWithMessage(400, "代码逻辑错误", (*context2.Context)(c.Ctx))
+		return
 	}
 	var TpProductService services.TpProductService
 	TpProduct := models.TpProduct{
-		Id: TpProductIdValidate.Id,
+		Id:       reqData.Id,
+		TenantId: tenantId,
 	}
 	rsp_err := TpProductService.DeleteTpProduct(TpProduct)
 	if rsp_err == nil {
-		response.SuccessWithMessage(200, "success", (*context2.Context)(TpProductController.Ctx))
+		response.SuccessWithMessage(200, "success", (*context2.Context)(c.Ctx))
 	} else {
 		var err string
 		isTrue := strings.Contains(rsp_err.Error(), "23503")
@@ -176,6 +204,6 @@ func (TpProductController *TpProductController) Delete() {
 		} else {
 			err = rsp_err.Error()
 		}
-		response.SuccessWithMessage(400, err, (*context2.Context)(TpProductController.Ctx))
+		response.SuccessWithMessage(400, err, (*context2.Context)(c.Ctx))
 	}
 }
